Moves everything related to MapConfigProvider to its own package
