+++ lib/windshaft/backends/map.js
@@ -5,7 +5,7 @@ var queue = require('queue-async');
 var step = require('step');
 var debug = require('debug')('windshaft:backend:map');
 
-var DummyMapConfigProvider = require('../models/dummy_mapconfig_provider');
+var DummyMapConfigProvider = require('../models/providers/dummy_mapconfig_provider');
 var Timer = require('../stats/timer');
 
 /**
