+++ lib/windshaft/backends/map_validator.js
@@ -1,7 +1,7 @@
 var queue = require('queue-async');
 var _ = require('underscore');
 
-var MapConfigProviderProxy = require('../models/mapconfig_provider_proxy');
+var MapConfigProviderProxy = require('../models/providers/mapconfig_provider_proxy');
 
 /**
  * @param {TileBackend} tileBackend
