+++ lib/windshaft/index.js
@@ -11,7 +11,7 @@ module.exports = {
         Datasource: require('./models/datasource'),
         MapConfig: require('./models/mapconfig'),
         provider: {
-            MapStoreMapConfig: require('./models/mapstore_mapconfig_provider')
+            MapStoreMapConfig: require('./models/providers/mapstore_mapconfig_provider')
         }
     },
     storage: {
