+++ lib/windshaft/models/providers/dummy_mapconfig_provider.js
