+++ lib/windshaft/models/providers/mapconfig_provider_proxy.js
