+++ lib/windshaft/models/providers/mapstore_mapconfig_provider.js
