+++ test/integration/static_maps.js
@@ -5,7 +5,7 @@ var crypto = require('crypto');
 var http = require('http');
 var fs = require('fs');
 var windshaft = require('../../lib/windshaft');
-var DummyMapConfigProvider = require('../../lib/windshaft/models/dummy_mapconfig_provider');
+var DummyMapConfigProvider = require('../../lib/windshaft/models/providers/dummy_mapconfig_provider');
 
 var mapnik = require('mapnik');
 //var RedisPool = require('redis-mpool');
