+++ test/unit/render_cache.test.js
@@ -5,7 +5,7 @@ var assert = require('assert');
 var RendererCache = require('../../lib/windshaft/cache/renderer_cache');
 var MapStore = require('../../lib/windshaft/storages/mapstore');
 var MapConfig = require('../../lib/windshaft/models/mapconfig');
-var MapStoreMapConfigProvider = require('../../lib/windshaft/models/mapstore_mapconfig_provider');
+var MapStoreMapConfigProvider = require('../../lib/windshaft/models/providers/mapstore_mapconfig_provider');
 var RendererFactory = require('../../lib/windshaft/renderers/renderer_factory');
 var RedisPool = require('redis-mpool');
 var serverOptions = require('../support/server_options');
